Backend/Django/requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
billiard==4.2.0
boto3==1.26.89
botocore==1.29.89
bs4==0.0.2
celery==5.4.0
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.1.1
cron-descriptor==1.4.3
cryptography==39.0.2
cssselect==1.2.0
cycler==0.12.1
distro==1.8.0
Django==4.1.7
django-celery-beat==2.6.0
django-cors-headers==3.14.0
django-debug-toolbar==3.2.4
django-ebhealthcheck==2.0.2
django-filter==21.1
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-storages==1.13.2
django-timezone-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
dnspython==2.3.0
exceptiongroup==1.1.3
fake-useragent==1.5.1
feedparser==6.0.11
finnhub-python==2.4.19
fonttools==4.53.0
frozendict==2.4.4
frozenlist==1.4.0
gunicorn==20.1.0
h11==0.14.0
holidays==0.52
html5lib==1.1
httpcore==1.0.2
httpx==0.25.1
idna==3.4
importlib_metadata==7.1.0
importlib_resources==6.4.0
jmespath==1.0.1
jsonfield==3.1.0
kiwisolver==1.4.5
kombu==5.3.7
lxml==5.2.1
matplotlib==3.7.5
multidict==6.0.4
multitasking==0.0.11
mypy==1.10.0
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.28.0
packaging==24.1
pandas==1.5.3
pandas-ta==0.3.14b0
parse==1.20.1
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pyee==11.1.0
PyJWT==2.8.0
pymemcache==4.0.0
pymongo==4.3.3
pyodbc==5.0.1
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
python-crontab==3.2.0
python-decouple==3.8
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.7
requests==2.31.0
requests-html==0.10.0
s3transfer==0.6.0
scipy==1.10.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.0
swapper==1.3.0
tenacity==8.4.1
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.15
user-agent==0.1.10
vine==5.1.0
w3lib==2.1.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
wrapt==1.13.3
yahoo-fin==0.8.9.1
yarl==1.9.2
yfinance==0.2.40
zipp==3.18.1
